removed a warning
git-svn-id: https://pimpmobile.svn.sourceforge.net/svnroot/pimpmobile/trunk@393 3d5ecaf0-f903-0410-b953-c2c1a4d75763

+++ converter/load_xm.c
@@ -343,7 +343,7 @@ static BOOL load_instrument(FILE *fp, pimp_instrument *instr, struct pimp_sample
 			src_format = PIMP_SAMPLE_S16;
 			
 			prev = 0;
-			for (i = 0; i < sh->length / 2; ++i)
+			for (i = 0; i < (int)(sh->length / 2); ++i)
 			{
 				signed short data = 0;
 				fread(&data, 1, 2, fp);
